requirements.txt
# Core Application
streamlit==1.55.0
pandas==2.3.3
ortools==9.12.4544
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.10
tabulate==0.10.0
plotly==6.6.0

# AI Intelligence
google-genai==1.65.0

# Development & Testing
rich==14.3.3
